// agency-services-transunion/lib/interfaces/merge-report.interface.ts
export interface IBorrowerAddress {
  CreditAddress?: ICreditAddress;
  Dwelling?: ICodeRef;
  Origin?: ICodeRef;
  Ownership?: ICodeRef;
  Source?: ISource;
}
export interface ICreditAddress {
  city?: string;
  country?: string;
  county?: string;
  direction?: string;
  houseNumber?: string;
  postDirection?: string;
  stateCode?: string;
  streetName?: string;
  unit?: string;
  unparsedStreet?: string;
  postalCode?: string;
}